recipes-graphics/wayland/agl-compositor_git.bb: Bump SRCREV
This SRC_REV bump updates the gRPC proxy API to address some start-up
issues we're seeing when switching the qt homescreen shell client to
using the start-up + activation procedure rather than relying on the
wayland protocol to do that. A smaller fix for restarting the
compositor, as well as functionality to dynamically move windows with
it and scale windows was added in this update.

The recipe changes does include a version bump for the compositor as
well, making it easier to track down the compositor version.

In more detail, the following changes have been added since the previous
update:

- bdc551314dc5 meson.build: Bump version to 0.0.22
- be197db866b8 grpc-proxy: Use smaller waiting time
- 73e82a6f346d grpc-proxy: Start the gRPC server and waiting thread much sooner
- 53b470c60986 grpc-proxy: Fix missing args LOG()
- 0d2e33ded529 grpc-proxy/main-grpc: Don't fall trough assuming we have a wl_display
- 994d8a30b27f grpc-proxy: Don't attempt to connect if there's no compositor running
- d8e72099ecbc grpc-proxy: Terminate thread when we're exiting
- ca537df2c529 grpc-proxy: Add dynamic scale of floating windows
- 2edba91d3eab grpc-proxy: Add dynamic floating window movement

Change-Id: Ic07b11e6334f14f60a1d71e48347a3d66f07b823
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/29405
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>