Remove remaining "display" group usage
Update the meta-rcar-gen3 driver specific bbappends to use the
"video" group instead of "display", and remove the weston-init
bbappend that was adding the "display" group when using it with
weston.  The group is no longer used in meta-rcar-gen3 or in any
of the AGL layers since the rework done for SPEC-4714; these
uses were missed at that time, and were causing permission issues
with the PVR driver initialization scripts.

Bug-AGL: SPEC-4979

Change-Id: I2efc8ea6e83b86e59def8a8333c8cc3064f8a7dd
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/29456
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account