Enhance DEFAULTTUNE mechanism
This commit enhances the way we select the DEFAULTTUNE.

DEFAULTTUNE in AGL is meant to be set to either
corei7-64 for x86-64
aarch64 for 64bit ARM
armv7vethf-neon-vfpv4 for 32bit ARM (where applicable)

The reason is that we do want to work towards a binary distro and
also maximize sstate reuse between platforms. Each and every board
having different settings is counter-productive here.

The original setting was done in SPEC-491. See this presentation:
https://docs.google.com/presentation/d/1_VSGfheItufCCP_rbGBDg3-2miCyh02-yZva6m8cyUE

The enhancement is that we introduce a variable 'AGL_DEFAULTTUNE'
that a user could override if he really needs to change the defaults.
Also we introduce OVERRIDES to help support this mechanism
(agldefaulttune, forcedefaulttune).

Bug-AGL: SPEC-491
Bug-AGL: SPEC-5014
Bug-AGL: SPEC-5016

Change-Id: Ia8270b6195116151c27d8da427f0456f38ac95b6
Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/29541
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account